add Dotty to build
including adding Scala 3 version of ShouldNotTypecheck
