+++ .travis.yml
@@ -6,6 +6,7 @@ language: scala
 
 scala:
   - 2.13.3
+  - 0.27.0-RC1
 
 env:
   - ADOPTOPENJDK=8
