+++ junit/src/test/scala-2/scala/SerializationStabilityTest.scala
@@ -18,72 +18,12 @@ import org.junit.Test
 
 // This test is self-modifying when run as follows:
 //
-//    junit/testOnly scala.SerializationStabilityTest -- -Doverwrite.source=src/test/scala/scala/SerializationStabilityTest.scala
+//    junit/testOnly scala.SerializationStabilityTest -- -Doverwrite.source=src/test/scala-2/scala/SerializationStabilityTest.scala
 //
 // Use this to re-establish a baseline for serialization compatibility.
 
 // based on run/t8549.scala partest
-object SerializationStability extends App {
-
-  val overwrite: Option[Path] =
-    sys.props.get("overwrite.source")
-      .map(s => Paths.get(s).toAbsolutePath)
-
-  def serialize(o: AnyRef): String = {
-    val bos = new java.io.ByteArrayOutputStream()
-    val out = new java.io.ObjectOutputStream(bos)
-    out.writeObject(o)
-    out.flush()
-    printBase64Binary(bos.toByteArray())
-  }
-
-  def amend(path: Path)(f: String => String): Unit = {
-    val old = new String(java.nio.file.Files.readAllBytes(path))
-    Files.write(path, f(old).getBytes)
-  }
-
-  def quote(s: String) = List("\"", s, "\"").mkString
-
-  def patch(path: Path, line: Int, prevResult: String, result: String): Unit = {
-    amend(path) {
-      content =>
-        content.linesIterator.toList.zipWithIndex.map {
-          case (content, i) if i == line - 1 =>
-            val newContent = content.replace(quote(prevResult), quote(result))
-            if (newContent != content)
-              println(s"- $content\n+ $newContent\n")
-            newContent
-          case (content, _) => content
-        }.mkString("\n")
-    }
-  }
-
-  def updateComment(path: Path): Unit = {
-    val timestamp = {
-      import java.text.SimpleDateFormat
-      val sdf = new SimpleDateFormat("yyyyMMdd-HH:mm:ss")
-      sdf.format(new java.util.Date)
-    }
-    val newComment = s"  // Generated on $timestamp with Scala ${scala.util.Properties.versionString})"
-    amend(path) {
-      content =>
-        content.linesIterator.toList.map {
-          f => f.replaceAll("""^ +// Generated on.*""", newComment)
-        }.mkString("\n")
-    }
-  }
-
-  def deserialize(string: String): AnyRef = {
-    val bis = new java.io.ByteArrayInputStream(parseBase64Binary(string))
-    val in = new java.io.ObjectInputStream(bis)
-    in.readObject()
-  }
-
-  def checkRoundTrip[T <: AnyRef](instance: T)(f: T => AnyRef): Unit = {
-    val result = serialize(instance)
-    val reconstituted = deserialize(result).asInstanceOf[T]
-    assert(f(instance) == f(reconstituted), (f(instance), f(reconstituted)))
-  }
+object SerializationStability extends App with SerializationStabilityBase {
 
   def check[T <: AnyRef](instance: => T)(prevResult: String, f: T => AnyRef = (x: T) => x): Unit = {
     val result = serialize(instance)
@@ -98,7 +38,7 @@ object SerializationStability extends App {
     }
   }
 
-  // Generated on 20190328-12:57:02 with Scala version 2.13.0-pre-59975bb)
+  // Generated on 20201015-13:14:45 with Scala version 2.13.3)
   overwrite.foreach(updateComment)
 
   // check(new collection.concurrent.TrieMap[Any, Any]())( "rO0ABXNyACNzY2FsYS5jb2xsZWN0aW9uLmNvbmN1cnJlbnQuVHJpZU1hcKckxpgOIYHPAwAETAALZXF1YWxpdHlvYmp0ABJMc2NhbGEvbWF0aC9FcXVpdjtMAApoYXNoaW5nb2JqdAAcTHNjYWxhL3V0aWwvaGFzaGluZy9IYXNoaW5nO0wABHJvb3R0ABJMamF2YS9sYW5nL09iamVjdDtMAAtyb290dXBkYXRlcnQAOUxqYXZhL3V0aWwvY29uY3VycmVudC9hdG9taWMvQXRvbWljUmVmZXJlbmNlRmllbGRVcGRhdGVyO3hwc3IAMnNjYWxhLmNvbGxlY3Rpb24uY29uY3VycmVudC5UcmllTWFwJE1hbmdsZWRIYXNoaW5nhTBoJQ/mgb0CAAB4cHNyABhzY2FsYS5tYXRoLkVxdWl2JCRhbm9uJDLBbyx4dy/qGwIAAHhwc3IANHNjYWxhLmNvbGxlY3Rpb24uY29uY3VycmVudC5UcmllTWFwU2VyaWFsaXphdGlvbkVuZCSbjdgbbGCt2gIAAHhweA==")
