+++ project/plugins.sbt
@@ -1,1 +1,3 @@
 addSbtPlugin("org.scala-lang.modules" % "sbt-scala-module" % "2.2.3")
+addSbtPlugin("ch.epfl.lamp" % "sbt-dotty" % "0.4.2")
+
